Sudoku solver
state function error narrowed down to the drawing of the box framework: if the box framework is assigned to a written out value, the function works, but if the framework if generated via code, it doesn't. The weird thing is that I can assert that one and the other are identical.
I don't know what's going on.